Add input functionality to prog.c
Hey Nikhil,

Your program helped me get started solving this problem, because I was having trouble with that. 

While looking at you program I noticed that part of the problem states that "Function fgetAnswers can also be called to input a student's answers," p. 443 in Hanly and Koffman, 7th Ed.

I added this functionality using the 'fopen' append option, with the "a" in place of the "r" or "w", and I thought it might be fun to see if you would add it in.

Please do!

Thank you,

Ryan